+++ packages/engine/src/lib/handler/piece-executor.ts
@@ -56,7 +56,7 @@ const executeAction: ActionHandler<PieceAction> = async ({ action, executionStat
 
         const { processedInput, errors } = await propsProcessor.applyProcessorsAndValidators(resolvedInput, pieceAction.props, piece.auth, pieceAction.requireAuth)
         if (Object.keys(errors).length > 0) {
-            throw new Error(JSON.stringify(errors))
+            throw new Error(JSON.stringify(errors, null, 2))
         }
 
         const hookResponse: HookResponse = {
