+++ packages/engine/src/lib/helper/trigger-helper.ts
@@ -248,7 +248,7 @@ async function prepareTriggerExecution({ pieceName, pieceVersion, triggerName, i
     const { processedInput, errors } = await propsProcessor.applyProcessorsAndValidators(resolvedInput, pieceTrigger.props, piece.auth, pieceTrigger.requireAuth)
 
     if (Object.keys(errors).length > 0) {
-        throw new Error(JSON.stringify(errors))
+        throw new Error(JSON.stringify(errors, null, 2))
     }
 
     return { piece, pieceTrigger, processedInput }
